pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "diffusion-prune"
version = "0.1.0"
description = "Pruning and quantization evaluation for diffusion language models"
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = { text = "Apache-2.0" }
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "torch>=2.0.0",
    "transformers>=5.0.0",
    "accelerate>=1.2.0",
    "lm-eval>=0.4.0",
    "hydra-core>=1.3.0",
    "omegaconf>=2.3.0",
    "rich>=13.0.0",
    "numpy>=1.24.0",
    "datasets>=2.14.0",
    "hydra_colorlog>=1.2.0",
    "hydra-submitit-launcher>=1.2.0",
    "typer>=0.12.0",
    "submitit>=1.5.0",
    "hf_transfer>=0.1.9",
    "matplotlib>=3.7.0",
    "vllm>=0.6.0",
]

[project.urls]
Repository = "https://github.com/Conzel/super-outlier-dlm"
Issues = "https://github.com/Conzel/super-outlier-dlm/issues"

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
    "pre-commit>=3.5.0",
    "pytest>=7.4.0",
    "ipykernel>=6.25.0",
    "jupyter>=1.0.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
]

[tool.ruff]
line-length = 100
target-version = "py311"
extend-exclude = ["src/diffusion_prune/evaluation/fast_dllm", "src/diffusion_prune/model", "experiments"]

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "UP",
]
ignore = [
    "E501",
    "B008",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.ruff.lint.isort]
known-first-party = ["diffusion_prune"]
